Move common schema statements methods to `mysql/schema_statements.rb`
